Don't create readonly container descriptors
This way, we'll be able to use the container descriptor's presence as
a means to indicate rw-ness, which is a simplification over checking
both that *and* then checking the rw flag. This simplification will
also aid us in moving what started out as the Rakudo Scalar container
spec into the MoarVM core, since then it won't need to be coupled to
the nested Scalar/Descriptor structure.

This doesn't yet do the final step of switching the check to look at
that; it just removes every place I could track down where we created
a readonly container descriptor in preparation. In theory, that means
the rest will be simple.

We lose one spectest, which checked for a variable name in some errors.
This isn't a great loss, since we could do that analysis and report on
it at compile time anyway, and that will be a reliable check, not an
unreliable one.