First small step for the assignment spesh plugin
So far, it just specializes the case where we have a simple assign
to do (no whence, no Nil, typecheck passes). It sets guards and then
turns this all into a bindattr, which spesh can then further lower to
a pointer write guarded with the GC write barrier.

No worthwhile measurements available yet, due to missing guard JIT and
the fact that the generated code still has a good bit of cruft that we
could do with eliminating.